requirements.txt
Flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
gunicorn>=21.2.0
deepface>=0.0.79
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
tensorflow>=2.13.0
pandas>=2.0.0





